Add option to read configuration from stdin
If configuration filename is a hyphen "-", we attempt to parse
the input from stdin. This makes it possible to modify configuration files
with linux command line tools and pipe the results to (for example)
"anemoi-dataset create". Both linux pipes and input redirection is supported.

cat config.yaml | anemoi-datasets create - test.zarr
anemoi-datasets create - test.zarr < config.yaml

We attempt to read the contents using the supported parsers in the following
order: yaml, json, toml.

If all parsers fail, an exception is raised.