refactor: use fixture for temp dir handling for test data (#166)
## Description
[Only to be merged when also both PRs linked below are reviewed/approved
since this PR requires changes in how the downloading utilities are used
downstream.]

This PR updates the handling of temporary directories in the testing
utilities. It changes the downloading functions into fixtures that use
pytest's tmp_path_factory fixture. The implemented caching behaviour (to
only download the same datasets once when used in multiple tests)
remains unchanged.

However, since the functions are replaced with fixtures they need to be
invoked as fixtures in the consuming repositories. The corresponding PRs
are these:
- for the integration tests in training:
https://github.com/ecmwf/anemoi-core/pull/372
- for tests in datasets: https://github.com/ecmwf/anemoi-core/pull/372

Together with the changes of this PR, the integration tests on the
branch in training as well as the unit tests on the branch in datasets
pass locally.

## What problem does this change solve?
More robust handling of temporary directories in testing utilities.

## What issue or task does this change relate to?
This would address the first point (handling of tmp dirs) of [this
issue](https://github.com/ecmwf/anemoi-utils/issues/147). The other
parts (like optional not-caching etc.) could be addressed separately, or
addressed by the exposed fixture temporary_directory_for_test_data as
done in the PR in core.

Compare also https://github.com/ecmwf/anemoi-datasets/issues/308

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***