fix: refactor code for casting dotdicts and apply this in getitem and setitem methods (#169)
## Description

This PR improves the behavior of the DotDict class to ensure consistent
recursive casting of nested dictionaries.

Previously, DotDict did not override `__setitem__`, which meant that
setting values like:

    a = DotDict({'a': 1})
    a['b'] = {'c': 'd'}

would result in a.b returning a plain dict instead of a DotDict.

Additionally, `__setattr__` only converted top-level dict values, but
did not handle nested structures such as lists or tuples of
dictionaries. For example, assigning a list of dicts would store the
list as-is, breaking the recursive dot-access pattern.
To fix these problems
1)  I implemented a `__setitem__` function on `DotDict` 
2) I created a method `to_dotdict` with code extracted from `__init__`
which is used to cast values passed to setitem/setattribute.

This ensures that both initialization and later updates maintain the
expected dot-access behavior throughout nested structures.

I also implemented tests to check that the changes works as intended. 


## What problem does this change solve?
Unexpected behavior of DotDict. 

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Florian Pinault <floriankrb@users.noreply.github.com>
Co-authored-by: b8raoult <53792887+b8raoult@users.noreply.github.com>