feat: NoAuth for AML mlflow Logging (#200)
## Description

This is a support PR for ecmwf/anemoi-core#469. My basic understanding
is that the authentication processes are applicable to AML mlflow
logging, since lots of credentials are required to setup the workspace
in the first place. So, this creates a PassiveAuth authenticator to
allow the AnemoiAzureMlflowLogger to be a child class of the
AnemoiMlflowLogger parent class, just silently ignoring all the
TokenAuth steps.

## What problem does this change solve?

New feature: supports AML mlflow logging.

## What issue or task does this change relate to?

None open but happy to open one if necessary.

##  Additional notes ##

See ecmwf/anemoi-core#469 discussion

---

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)